requirements.txt
#Adafruit_PCA9685 # For Servo interfacing from a Raspberry Pi
#RPi.GPIO # for GPIO interface on a Raspbery Pi
pyttsx3 #for text to speech
pywin32 # for text to speech
#smtplib #for speech to text
SpeechRecognition #for speech recognition 
pydub #for speech-to-text
wikipedia #for wikipedia searches
wolframalpha #for wolframalpha searches

#keyboard #for keyboard inputs
#pynput #for keyboard and mouse inputs

beautifulsoup4 #for google searches
python-dotenv #to open info out of a .env
google # for google searches

PyAudio #if this doesnt work on windows, unhash the next line and re-has this one:
#if you get errors from ^The pip install^:
#"https://www.lfd.uci.edu/~gohlke/pythonlibs/#pyaudio" - go here and download the appropriete whl for your system
# then use the following command:
# pip install "Path/To/File.whl"




import webbrowser #for opening web pages
import smtplib #for emailing
import random # to pick random

import wikipedia #to look stuff up
import sys # to use the system
import time #so you can wait
from googlesearch import search # to search google for things